Split Deprecation Objects for easier seperation of concerns
